Move team table generation from victory to teams
